Refactor rename model directory into models folder
