+++ src/mocks/response/archiveEntryResponse.ts
@@ -1,4 +1,4 @@
-import { ArchiveEntry } from '../../model/archiveEntry';
+import { ArchiveEntry } from '../../models/archiveEntry';
 import { Response, ResponseOptions} from '@angular/http';
 
 export class ArchiveEntryResponse extends Response {
