+++ src/mocks/response/entriesResponse.ts
@@ -1,4 +1,4 @@
-import { EntriesPoint } from '../../model/entries-point';
+import { EntriesPoint } from '../../models/entries-point';
 import { Response, ResponseOptions} from '@angular/http';
 
 export class EntriesResponse extends Response {
