+++ src/models/archiveTableEntry.ts
