+++ src/models/entries-point.ts
