+++ src/models/entry-point.ts
