+++ src/models/imageEntry.ts
