+++ src/models/info.ts
