+++ src/models/link.ts
