+++ src/providers/token-service.spec.ts
@@ -3,7 +3,7 @@ import { Http, HttpModule, BaseRequestOptions } from '@angular/http';
 import { MockImsBackend } from '../mocks/mock-ims-backend';
 import { TokenService } from './token-service';
 import { ImsService } from './ims-service';
-import { Token } from '../model/token';
+import { Token } from '../models/token';
 
 describe('Provider: TokenService', () => {
